requirements.txt
pdfplumber
pandas
openpyxl
jsonlines
python-dotenv
requests